requirements.txt
google-adk[a2a]
uvicorn
python-dotenv